Rewritten mkhlf in Perl; removed awk dependency

+++ far2l/src/console/interf.cpp
@@ -786,7 +786,7 @@ void ClearScreen(int Color)
 	Console.SetTextAttributes(Color);
 }
 
-DWORD64 GetColor()
+int GetColor()
 {
 	return (CurColor);
 }
